+++ include/chemfiles/capi/selection.h
@@ -76,20 +76,6 @@ CHFL_EXPORT chfl_status chfl_selection_evaluate(
     CHFL_SELECTION* selection, const CHFL_FRAME* frame, uint64_t* n_matches
 );
 
-/// Maximal size for a selection match
-#define CHFL_MAX_SELECTION_SIZE 4
-
-/// A `chfl_match` is a set of atomic indexes matching a given selection. The
-/// size of a match depends on the associated selection, and can vary from 1 to
-/// `CHFL_MAX_SELECTION_SIZE`.
-typedef struct {  // NOLINT: this is both a C and C++ file
-    /// The actual size of the match. Elements in `atoms` are significant up
-    /// to this value, and filled with `(uint64_t)-1` for all the other values.
-    uint64_t size;
-    /// Atomic indexes matching the associated selection
-    uint64_t atoms[CHFL_MAX_SELECTION_SIZE];
-} chfl_match;
-
 /// Get the matches for a `selection` after a call to `chfl_selection_evaluate`,
 /// in `matches`.
 ///