+++ src/capi/atom.cpp
@@ -6,9 +6,11 @@
 #include <string>
 
 #include "chemfiles/capi/types.h"
-#include "chemfiles/capi/atom.h"
+#include "chemfiles/capi/misc.h"
 #include "chemfiles/capi/utils.hpp"
 #include "chemfiles/capi/shared_allocator.hpp"
+
+#include "chemfiles/capi/atom.h"
 
 #include "chemfiles/Atom.hpp"
 #include "chemfiles/Frame.hpp"