+++ src/capi/frame.cpp
@@ -5,9 +5,11 @@
 #include <string>
 
 #include "chemfiles/capi/types.h"
-#include "chemfiles/capi/frame.h"
+#include "chemfiles/capi/misc.h"
 #include "chemfiles/capi/utils.hpp"
 #include "chemfiles/capi/shared_allocator.hpp"
+
+#include "chemfiles/capi/frame.h"
 
 #include "chemfiles/Frame.hpp"
 #include "chemfiles/error_fmt.hpp"